feat: update Candle model to use FloatField for volume and enhance historical data fetching for asset classes

+++ backend/apps/core/models.py
@@ -198,7 +198,8 @@ class Candle(models.Model):
     high = models.FloatField()
     low = models.FloatField()
     close = models.FloatField()
-    volume = models.BigIntegerField()
+    # for crypto, volume can be fractional
+    volume = models.FloatField()
 
     # Additional Alpaca fields
     trade_count = models.IntegerField(blank=True, null=True)  # 'n' field
